+++ WarpLib/TiltSeries/TiltSeries.ReconstructParticleSeries.cs
@@ -146,7 +146,7 @@ public partial class TiltSeries
             ImagesFT = GetImagesForOneParticle(options, TiltData, SizeSub, ParticlePositions, PlanForwParticle, -1, 0, false, Images, ImagesFT);
             GetCTFsForOneParticle(options, ParticlePositions, CTFCoords, null, false, false, false, CTFs);
 
-            if (!options.ExtractRaw)
+            if (!options.DontPremultiply)
             {
                 ImagesFT.Multiply(CTFs);
                 ImagesFT.Multiply(RelionWeights);